requirements.txt
pandas>=1.3.0
celery>=5.2.0
kombu>=5.2.0
psycopg2-binary>=2.9.0
scikit-learn>=1.0.0
numpy>=1.21.0
tensorflow==2.16.1